unidossis_pqrs/requirements.txt
# ─── Framework principal ─────────────────────────────────
Django==6.0.4
asgiref==3.11.1
sqlparse==0.5.5

# ─── Base de datos ────────────────────────────────────────
# psycopg2-binary   # Descomentar al usar PostgreSQL en producción (Railway)

# ─── Archivos y exportación ──────────────────────────────
openpyxl==3.1.5
et_xmlfile==2.0.0
pillow==12.2.0

# ─── Análisis de datos ───────────────────────────────────
pandas==3.0.2
numpy==2.4.8
python-dateutil==2.9.0.post0
six==1.17.0
pytz==2025.2

# ─── IA / Google Generative AI ───────────────────────────
google-genai==1.72.0
google-generativeai==0.8.6
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
proto-plus==1.27.2
protobuf==5.29.6

# ─── HTTP / Red ───────────────────────────────────────────
requests==2.33.1
httpx==0.28.1
httpcore==1.0.9
httplib2==0.31.2
h11==0.16.0
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
urllib3==2.6.3
uritemplate==4.2.0

# ─── Criptografía / Seguridad ────────────────────────────
cryptography==46.0.7
cffi==2.0.0
pycparser==3.0
pyasn1==0.6.3
pyasn1_modules==0.4.2

# ─── Validación y tipos ──────────────────────────────────
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# ─── Utilidades ───────────────────────────────────────────
colorama==0.4.6
distro==1.9.0
tqdm==4.67.3
tenacity==9.1.4
anyio==4.13.0
sniffio==1.3.1
pyparsing==3.3.2
websockets==16.0
tzdata==2026.1

# ─── Servidor de producción (Railway) ────────────────────────
gunicorn==25.3.0
whitenoise==6.12.0
dj-database-url==3.1.2
psycopg2-binary==2.9.11
packaging==26.1
